add json & custom support
