+++ examples/json.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width" />
+    <title></title>
+    <script type="module" src="../docs/module.js"></script>
+  </head>
+  <body>
+    <script type="text/yaml">
+      title: 中国地方の自然と産業
+      accessToken: pk.eyJ1IjoiZ2VvcmVwdWJsaWMiLCJhIjoiY2tvMm53aWsxMTFiajJycGdyOXR4MW9iMiJ9.h7aSUinQuXd8EFRSWJvBPQ
+      style: https://optgeo.github.io/b3p/style.json
+      chapters: ./test.json
+      footer: >-
+        <p>このプロジェクトは、
+        <a href='https://github.com/optgeo'>Adopt Geodata プロジェクト</a>の一環です。</p>
+    </script>
+  </body>
+</html>
