+++ examples/test.json
@@ -1,0 +1,1 @@
+[{"title":"岡山平野","description":"岡山県にある平野です。","hash":"11.19/34.5801/133.8456/63.2/59"},{"title":"鳥取砂丘","description":"鳥取県にある砂丘です。","hash":"14.9/35.54016/134.22845/153.6/0"}]
